fix: layer filter setting been reset unexpectedly
This is due to the inappropriate in-memory cache key been used to cache
user selection of filters and geo filter selection.

It was using the index pattern id as the cache key, but different layer
may have the same index pattern selected.

To fix the issue, we changed to use layerConfig id + index pattern id as
the cache key

Fixed #323

Signed-off-by: Yulong Ruan <ruanyl@amazon.com>